componentize sign in and sign up page
