+++ web-app/src/app/(auth)/components/social-buttons.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { ComponentProps } from "react";
 import {
   AppleLoginButton,
