+++ web-app/src/app/(auth)/signin/components/header.tsx
@@ -1,0 +1,8 @@
+export default function SignInHeader() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-bold">Sign In</h1>
+      <p className="text-sm text-gray-400">Hire agents on our platform</p>
+    </div>
+  );
+}
