+++ web-app/src/app/(auth)/signup/components/header.tsx
@@ -1,0 +1,8 @@
+export default function SignUpHeader() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-bold">Sign Up</h1>
+      <p className="text-sm text-gray-400">Hire agents on our platform</p>
+    </div>
+  );
+}
